# parser: use `memchr` to speed-up `skip_till()`