@GuillaumeGomez 2024-07-25
PR looks good. Just a few things I need to confirm: you mentioned that `memchr` only works on ASCII characters. However, we can overload syntax (like comment start/end for example). Nothing is preventing users to use non-ASCII characters for that. Will it still work in this case? If yes, please add tests to ensure it.