@Kijewski 2024-07-25
> you mentioned that memchr only works on ASCII characters

`memchr` works on any `&[u8]`. It does not know or care what the input it.

But we have to care, because we must not split a `&str`. The first byte of a UTF-8 encoded code point is either:
- `0b0xxx_xxxx` (ASCII, 1 byte),
- `0b110x_xxxx` (BMP, 2 bytes),
- `0b1110_xxxx` (3 bytes) or
- `0b1111_0xxx` (4 bytes),
- but never `0b10xx_xxxx`.

Knowing that, we know that the result of `memchr` will never point into a UTF-8 sequence, only at the start of one, because there is no way that `memchr` could have searched for a byte inside a UTF-8 sequence.