@Kijewski 2024-07-28
Rebas on the current HEAD. Some other PR did make the parsing faster in the meantime, too, so I re-did the benchmark. Updated the initial comment. Still a good speed-up.